+++ standalone-metastore/metastore-server/src/main/java/org/apache/hadoop/hive/metastore/metastore/impl/ColStatsStoreImpl.java
@@ -1217,6 +1217,4 @@ public class ColStatsStoreImpl extends RawStoreAware implements ColStatsStore {
       }
     }
   }
-
-
 }
